requirements.txt
python-docx
pytesseract
pillow
pymupdf
fastapi
uvicorn
python-dotenv
openai
python-multipart